Revert commit c0547c56dafb ('Add configure option to disable DES')
AUTH_DES is part of the legacy SunRPC API in glibc, and therefore
is not optional.

Commit c0547c56dafb says making AUTH_DES optional "is needed for
musl libc." Removing AUTH_DES breaks legacy code, thus some other
fix must be found for this issue. libtirpc can provide some or all
of the missing functionality if the local C library does not have
it, perhaps.

Reported-by: Thorsten Kukuk <kukuk@suse.de>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>