Remove vestiges of AUTH_KERB
AUTH_KERB is described only briefly in the ONC+ Developer's Guide.

However, xdr_authkerb_cred(), for example, is declared in Linux's
tirpc/rpc/auth.h, but does not appear in Solaris 11's RPC headers.

It is not provided by glibc, so there isn't any backward
compatibility requirement.

The libtirpc implementation of AUTH_KERB here on Linux and in
FreeBSD seems to be incomplete and inoperative.

Seems like the scattered pieces of it can be removed safely.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>