Handle gss auth reference counting internally
The previous implementation of auth_get and auth_put used atomic
builtins to allow thread-safe updating of a reference count for the AUTH
object. This causes libtirpc not to build on a number of
non-mainstream architectures which don't support the atomic builtins.

In addition it exposed the feature as macros through a header file,
rather than through the ABI, preventing the underlying implementation
from being updated.

This patch removes the API calls and handles the reference counting/
deallocation of the AUTH object internally.

Signed-off-by: Brendan Heading <brendanheading@gmail.com>
Signed-off-by: Steve Dickson <steved@redhat.com>