libtirpc: replace array with list for per-fd locks
Currently per-fd locks for the clients are pre-allocated up to
the soft limit of maximum allowed open file desciptors per process as
defined in __rpc_dtbsize():

	if (getrlimit(RLIMIT_NOFILE, &rl) == 0) {
                return (tbsize = (int)rl.rlim_cur);
        }

This limit can be arbitrarily large for any given process resulting in
unreasonable memory allocation. For example, for systemd PID1 process
this limit is set to 1073741816 since version 240. systemd is an
indirect user of this library as it fetches information about users,
groups, etc...

This patch proposes a list implementation of per-fd locks based on glibc
doubly linked lists. It also includes support for a fixed array based
pre-allocation up to a compile-time defined limit of locks for
equivalence to the previous implementation.

Signed-off-by: Steve Dickson <steved@redhat.com>